// app/api/types/legislationDocuments.ts
export interface LegislativeDocument {
    name: string
    shortFriendlyName: string
    biennium: string
    longFriendlyName: string
    description: string
    type: string
    class: string
    htmUrl: string
    htmCreateDate: string
    htmLastModifiedDate: string
    pdfUrl: string
    pdfCreateDate: string
    pdfLastModifiedDate: string
    billId: number
}